Second Phase
It kills the two hard-coded processes now but occasionally throws an
exception about "access". It looks like it is because it kills the
process but then tries to kill it again before a new array is created. I
don't want to add any delays (sleeps) so I'm working on making a list so
I can pop processes as they are terminated.